Converted all the id's of Entity into a single layer.
Actually that's not the Problemm of the last commit .
What i implemented is that i already has positions of the entity and next what i has to do is to just paste it in a layer this also removed the headache of the 3rd layer which i initially used to store the id's of moving entities only...
Hurrah......

+++ game/simulation/generate_map.py
@@ -12,7 +12,7 @@ def generate_map(width, height, seed=None):
     np.random.seed(seed)
     layer_0=np.zeros((width, height), dtype=np.int8) # for tile layer
     #This is the fixed layer later will be replaced by procedural generation.
-    layer_1=np.zeros((width, height), dtype=np.int8)
+    layer_1=np.zeros((width, height), dtype=np.int16)
     
     rng = np.random.default_rng()
